chore: p0 hygiene — barrel trim, micro-file merges, runtime word
Derived on post-sweep main per #453's rules: 1 re-export-only barrel
(event/index.ts) inlined into the root barrel with a vocab-baseline
shrink (30→29 namespaces); 20 sub-30-LOC single-importer micro-files
merged into their importers (2 skipped: plan.ts is baseline-referenced,
github/webhook.ts separation is load-bearing); cheap standalone-runtime
string fixes applied, structural renames listed as P3 carry-overs.
