--- apps/server/src/channel/discord/reconnect-backoff.ts
@@ -1,6 +1,0 @@
-export const FATAL_CLOSE_CODES = new Set([4004, 4010, 4011, 4012, 4013, 4014]);
-const MAX_BACKOFF_MS = 60_000;
-
-export function calculateBackoff(attempt: number): number {
-  return Math.min(1000 * 2 ** attempt, MAX_BACKOFF_MS) + Math.random() * 1000;
-}
