+++ apps/server/src/channel/discord/surface.ts
@@ -3,7 +3,6 @@ import { Operational, PolicyDecision } from "@openomni/protocol";
 import { Bus, SurfaceKey } from "@openomni/session";
 import { Dedupe } from "../../shared/dedupe";
 import { DiscordClient } from "./client";
-import { sendDiscordMessage } from "./formatter";
 import { DiscordGateway } from "./gateway";
 import { DiscordNormalizer } from "./normalizer";
 import type { DiscordMessage } from "./types";
@@ -161,5 +160,22 @@ export class DiscordAdapter implements Adapter.Surface {
     } finally {
       clearInterval(typingInterval);
     }
+  }
+}
+
+// merged from formatter.ts (#453 hygiene: sub-30-LOC single-importer)
+import { splitText } from "../../shared/chunk-text";
+import type { ChannelClient } from "../types";
+
+const DISCORD_MESSAGE_LIMIT = 2000;
+
+export async function sendDiscordMessage(
+  client: ChannelClient,
+  channelId: string,
+  message: Adapter.OutboundMessage,
+): Promise<void> {
+  if (!message.text) return;
+  for (const chunk of splitText(message.text, DISCORD_MESSAGE_LIMIT)) {
+    await client.send(channelId, chunk);
   }
 }
