--- packages/openomni/src/execution-runtime/tool/builtins/glob-prompt.ts
@@ -1,3 +1,0 @@
-export const GLOB_PROMPT = `Match files against a glob pattern (e.g. '**/*.ts') inside the workspace.
-Results are sorted newest-first by mtime and capped at 100 entries.
-Paths must stay within the workspace root; symlink escapes are rejected.`;
