--- packages/openomni/src/execution-runtime/tool/builtins/grep-prompt.ts
@@ -1,3 +1,0 @@
-export const GREP_PROMPT = `Search file contents for a regex pattern across the workspace.
-Use include (glob) to narrow the files searched and ignoreCase for case-insensitive matches.
-Returns up to 100 entries as {file, line, text}. Paths must stay within the workspace root.`;
