--- packages/openomni/src/execution-runtime/tool/builtins/write-prompt.ts
@@ -1,3 +1,0 @@
-export const WRITE_PROMPT = `Write content to a file inside the workspace.
-Creates parent directories as needed and overwrites any existing file at the path.
-The path must stay within the workspace root; symlink escapes are rejected.`;
