--- packages/openomni/src/execution-runtime/tool/executor-result.ts
@@ -1,10 +1,0 @@
-import type { Tool } from "@openomni/protocol";
-
-export function createErrorResult(call: Tool.Call, message: string): Tool.Result {
-  return {
-    id: crypto.randomUUID(),
-    toolCallId: call.id,
-    output: message,
-    isError: true,
-  };
-}
