+++ packages/openomni/src/ingress/handlers.ts
@@ -5,8 +5,8 @@ import {
   publishModeDetected,
   summarizeTarget,
 } from "./handler-events";
-import { buildExecutionRequest as buildRequest } from "./handler-request";
 import type { HandlerContext as IngressHandlerContext } from "./handler-types";
+import { extractPrompt } from "./handler-prompt";
 import {
   handleCancelRequest,
   handleWorkerDelivery,
@@ -26,7 +26,26 @@ export namespace IngressHandlers {
   export interface HandlerContext extends IngressHandlerContext {}
 
   export function buildExecutionRequest(ctx: HandlerContext): Execution.Request {
-    return buildRequest(ctx);
+    return {
+      runId: crypto.randomUUID(),
+      sessionId: ctx.sessionId,
+      mode: "direct",
+      prompt: extractPrompt(ctx.event.payload),
+      model: ctx.event.agent.model,
+      systemPrompt: ctx.event.agent.systemPrompt,
+      tools: ctx.event.agent.tools,
+      toolConfig: ctx.event.agent.toolConfig,
+      permissions: ctx.event.agent.permissions,
+      credentials: undefined,
+      budget: ctx.event.agent.budget,
+      workspaceRoot: ctx.event.agent.toolConfig?.workspaceRoot,
+      policyPlan: ctx.event.agent.policyPlan,
+      providerOptions: (ctx.event.agent as { providerOptions?: Record<string, unknown> })
+        .providerOptions,
+      traceId: ctx.traceContext?.traceId,
+      agentName:
+        typeof ctx.event.meta?.agentName === "string" ? ctx.event.meta.agentName : undefined,
+    };
   }
 
   export async function dispatchWritebackCommit(
