--- packages/openomni/src/ingress/middleware/ingress-authority-blacklist.ts
@@ -1,3 +1,0 @@
-export function blacklistReason(kind: string, value: string, reason: string | undefined): string {
-  return reason ?? `blacklist.${kind}.${value}`;
-}
