--- packages/protocol/src/event/index.ts
@@ -1,7 +1,0 @@
-export * from "./agent-execution.js";
-export * from "./ingress.js";
-export * from "./llm.js";
-export * from "./tool.js";
-export * from "./mcp.js";
-export * from "./operational.js";
-export * from "./policy.js";
