+++ packages/protocol/src/index.ts
@@ -6,7 +6,13 @@ export * from "./model/index.js";
 export * from "./run/index.js";
 export * from "./sink/index.js";
 export * from "./bus/index.js";
-export * from "./event/index.js";
+export * from "./event/agent-execution.js";
+export * from "./event/ingress.js";
+export * from "./event/llm.js";
+export * from "./event/tool.js";
+export * from "./event/mcp.js";
+export * from "./event/operational.js";
+export * from "./event/policy.js";
 export * from "./mcp/index.js";
 export * from "./adapter/index.js";
 export * from "./actor/index.js";
