--- packages/protocol/src/work-item/http.ts
@@ -1,9 +1,0 @@
-export function isHttpUrl(target: string): boolean {
-  try {
-    const protocol = new URL(target).protocol;
-    return protocol === "http:" || protocol === "https:";
-  } catch (error) {
-    if (error instanceof TypeError) return false;
-    throw error;
-  }
-}
