+++ packages/session/src/bus-persistence/event-query.ts
@@ -1,6 +1,5 @@
-import type { EventRecord, QueryOptions } from "./query-contracts.js";
+import { EventVisibility, type EventRecord, type QueryOptions } from "./query-contracts.js";
 import { buildEventFilters, limitClause } from "./event-query-filter.js";
-import { toEventRecord } from "./event-record-mapper.js";
 import { getDatabase } from "./query-database.js";
 import type { BusEventRow } from "./query-rows.js";
 
@@ -43,4 +42,20 @@ export function listForLlmReasoning(
     ...options,
     visibilityIn: ["llm_reason", "user_audit"],
   });
+}
+
+// merged from event-record-mapper.ts (#453 hygiene: sub-30-LOC single-importer)
+export function toEventRecord(row: BusEventRow): EventRecord {
+  return {
+    id: String(row.id),
+    sessionId: row.session_id,
+    runId: row.run_id ?? undefined,
+    eventType: row.event_type,
+    category: row.category,
+    visibility: EventVisibility.parse(row.visibility),
+    data: JSON.parse(row.data) as Record<string, unknown>,
+    traceId: row.trace_id,
+    durationMs: row.duration_ms ?? undefined,
+    timeCreated: row.time_created,
+  };
 }
