--- packages/session/src/work-item/completion-report.ts
@@ -1,25 +1,0 @@
-import { WorkItem } from "@openomni/protocol";
-
-export function verifyCompletionReport(
-  item: WorkItem.Info,
-  completionReport: WorkItem.CompletionReport,
-): WorkItem.CompletionReport {
-  const report = WorkItem.CompletionReport.parse(completionReport);
-  const evidenceById = new Map(item.evidence.map((evidence) => [evidence.id, evidence]));
-  const missing = report.claims.flatMap((claim) =>
-    claim.evidenceIds.filter((evidenceId) => !evidenceById.has(evidenceId)),
-  );
-  if (missing.length > 0) {
-    throw new Error(`completion report references missing evidence: ${missing.join(", ")}`);
-  }
-  const failed = report.claims.flatMap((claim) =>
-    claim.evidenceIds.filter((evidenceId) => {
-      const evidence = evidenceById.get(evidenceId);
-      return evidence?.passed === false || evidence?.readBack?.passed === false;
-    }),
-  );
-  if (failed.length > 0) {
-    throw new Error(`completion report references failed evidence: ${failed.join(", ")}`);
-  }
-  return report;
-}
