+++ packages/session/src/work-item/lifecycle.ts
@@ -1,7 +1,6 @@
 import { WorkItem } from "@openomni/protocol";
 import { Bus } from "../bus/index.js";
 import { Storage } from "../storage/storage.js";
-import { verifyCompletionReport } from "./completion-report.js";
 import { areWorkItemDependenciesMet } from "./dependency.js";
 import { mutate, mutateTimestamps, persistMutation } from "./mutation.js";
 import { recordWorkItemOutcome } from "./outcome.js";
@@ -135,4 +134,30 @@ export const areDependenciesMet = areWorkItemDependenciesMet;
 
 export async function retryStoredWorkItem(hash: string): Promise<WorkItem.Info | undefined> {
   return retryWorkItem(hash, Storage.get().workItem, persistMutation);
+}
+
+// merged from completion-report.ts (#453 hygiene: sub-30-LOC single-importer)
+
+export function verifyCompletionReport(
+  item: WorkItem.Info,
+  completionReport: WorkItem.CompletionReport,
+): WorkItem.CompletionReport {
+  const report = WorkItem.CompletionReport.parse(completionReport);
+  const evidenceById = new Map(item.evidence.map((evidence) => [evidence.id, evidence]));
+  const missing = report.claims.flatMap((claim) =>
+    claim.evidenceIds.filter((evidenceId) => !evidenceById.has(evidenceId)),
+  );
+  if (missing.length > 0) {
+    throw new Error(`completion report references missing evidence: ${missing.join(", ")}`);
+  }
+  const failed = report.claims.flatMap((claim) =>
+    claim.evidenceIds.filter((evidenceId) => {
+      const evidence = evidenceById.get(evidenceId);
+      return evidence?.passed === false || evidence?.readBack?.passed === false;
+    }),
+  );
+  if (failed.length > 0) {
+    throw new Error(`completion report references failed evidence: ${failed.join(", ")}`);
+  }
+  return report;
 }
