+++ script/conformance/lint-tools-baseline.json
@@ -11,7 +11,6 @@
       "cron",
       "dispatch",
       "error",
-      "event",
       "execution",
       "extension",
       "ingress",
